+++ index.html
@@ -3,9 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    
     <title>Test</title>
-    <script src="js/main.js"></script>
+    <script type="module" src="js/main.js"></script>
     <link rel="stylesheet" href="css/styles.css">
 </head>
 <body>
