+++ js/createDOM/clearDOM.js
@@ -1,0 +1,1 @@
+export default () => document.getElementById('app').innerHTML = '';
