+++ js/createDOM/showInfo.js
@@ -1,0 +1,9 @@
+export default (userDOM) => {
+    userDOM.style.display = 'block';
+
+    window.onclick = function(event) {
+        if (event.target == userDOM) {
+          userDOM.style.display = "none";
+        }
+    };
+}
